treewide: convert enums to ints before formatting
It is not well defined whether enums are signed or unsigne in the C
standard.  gcc treats them as unsigned, but other compilers are free
to do otherwise.  This causes problems when trying to compile this
codebase with "-Wformat-signedness".

Cast them to ints before formatting them to sidestep compiler
warnings.  These enums will not realistically exceed INT_MAX, so it is
safe to do so and expect the correct output.

This is another step towards getting this codebase ready to be
compiled with "-Wformat-signedness" turned on.

+++ builtin/bisect--helper.c
@@ -707,7 +707,7 @@ int cmd_bisect__helper(int argc, const char **argv, const char *prefix)
 		res = bisect_start(&terms, no_checkout, argv, argc);
 		break;
 	default:
-		return error("BUG: unknown subcommand '%d'", cmdmode);
+		return error("BUG: unknown subcommand '%d'", (int)cmdmode);
 	}
 	free_terms(&terms);
 	return !!res;
